CRUD Album

+++ src/main/java/com/api_musicas/service/ArtistaService.java
@@ -2,17 +2,18 @@ package com.api_musicas.service;
 
 import com.api_musicas.domain.ArtistaDTO;
 import com.api_musicas.mapper.ArtistaMapper;
+import com.api_musicas.model.AlbumModel;
 import com.api_musicas.model.ArtistaModel;
+import com.api_musicas.repository.AlbumRepository;
 import com.api_musicas.repository.ArtistaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import org.springframework.data.domain.Pageable;
 import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 
 import static com.api_musicas.util.Constantes.*;
 
@@ -25,6 +26,8 @@ public class ArtistaService {
     private ArtistaRepository repository;
     @Autowired
     private ImagemService imagemService;
+    @Autowired
+    private AlbumRepository albumRepository;
 
     public String save(ArtistaDTO dto) {
         ArtistaModel artistaModel = mapper.dtoToModel(dto);
@@ -72,9 +75,15 @@ public class ArtistaService {
         ArtistaModel model = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException(ARTISTA_N_ENCONTRADO));
 
-        repository.delete(model);
+        List<AlbumModel> albumModel = albumRepository.findByArtistaId(id);
 
-        return ARTISTA_DELETADO;
+        if(albumModel.isEmpty()){
+            repository.delete(model);
+
+            return ARTISTA_DELETADO;
+        }
+
+        return ERRO_ARTISTA_DELETADO;
 
     }
 }
